[fix]#2 コメントにて指摘された点の修正
- ディレクトリ構成の修正
- コンマの追加
- キーボード表示中、他の画面に触るとキーボードを閉じるよう修正
- 最後の入力欄より前では、キーボードの完了ボタンが「次へ」になるよう変更
- フォームの位置を固定するよう変更
