--- lib/screens/login_signup_screen/components/body.dart
@@ -1,119 +1,0 @@
-import 'package:flutter/material.dart';
-
-import 'package:dish/screens/login_signup_screen/components/social_icon_buttons.dart';
-import 'package:dish/screens/login_signup_screen/components/switch_tab_button.dart';
-import 'package:dish/screens/login_signup_screen/components/confirm_button.dart';
-import 'package:dish/screens/login_signup_screen/components/input_form.dart';
-import 'package:dish/configs/constant_colors.dart';
-
-class Body extends StatefulWidget {
-  @override
-  _BodyState createState() => _BodyState();
-}
-
-class _BodyState extends State<Body> {
-  String selectedTab = "login";
-  String email = "";
-  String password = "";
-  String confirmPassword = "";
-
-  void setEmail(String email) {
-    setState(() {
-      this.email = email;
-    });
-  }
-
-  void setPassword(String password) {
-    setState(() {
-      this.password = password;
-    });
-  }
-
-  void setConfirmPassword(String password) {
-    setState(() {
-      this.confirmPassword = password;
-    });
-  }
-
-  void switchTab(String selectedTab) {
-    setState(() {
-      this.selectedTab = selectedTab;
-      print(selectedTab);
-    });
-  }
-
-  void onPressedConfirmButton() async {
-    try {
-      if (selectedTab == "login") {
-        // ログイン処理を書く
-        print(selectedTab);
-      } else if (selectedTab == "signup") {
-        // 新規登録処理を書く
-        print(selectedTab);
-      }
-    } catch (e) {
-      print(e.toString());
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final String _backgroundImagePath = "assets/images/background.png";
-    return Container(
-      decoration: BoxDecoration(
-        image: DecorationImage(
-          image: AssetImage(_backgroundImagePath),
-          fit: BoxFit.cover,
-        ),
-      ),
-      child: Center(
-        child: Container(
-          height: 540,
-          margin: EdgeInsets.symmetric(horizontal: 35),
-          padding: EdgeInsets.symmetric(horizontal: 40, vertical: 56),
-          decoration: BoxDecoration(
-            color: Colors.white,
-            borderRadius: BorderRadius.circular(35),
-            boxShadow: [
-              BoxShadow(
-                color: Colors.black12,
-                blurRadius: 8,
-                spreadRadius: 4,
-                offset: Offset(0, 0),
-              )
-            ],
-          ),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              SwitchTabButton(switchTab: switchTab),
-              Spacer(flex: 2),
-              InputForm(
-                selectedTab: selectedTab,
-                setEmail: setEmail,
-                setPassword: setPassword,
-                setConfirmPassword: setConfirmPassword,
-              ),
-              Spacer(flex: 2),
-              ConfirmButton(
-                selectedTab: selectedTab,
-                press: onPressedConfirmButton,
-              ),
-              Spacer(flex: 1),
-              Text(
-                "OR",
-                style: TextStyle(
-                  color: AppColor.kPrimaryTextColor.withOpacity(0.6),
-                  fontSize: 14,
-                  letterSpacing: 2,
-                ),
-              ),
-              Spacer(flex: 1),
-              SocialIconButtons(),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
