--- lib/screens/login_signup_screen/login_signup_screen.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:dish/screens/login_signup_screen/components/body.dart';
-
-class LoginSignupPage extends StatefulWidget {
-  @override
-  _LoginSignupPageState createState() => _LoginSignupPageState();
-}
-
-class _LoginSignupPageState extends State<LoginSignupPage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Body(),
-    );
-  }
-}
